Move a pasta 'lib' para 'packtools/lib'
