+++ packtools/sps/models/sps_package.py
@@ -2,7 +2,7 @@ import logging
 import os
 
 from lxml import etree
-from lib import normalizer
+from packtools.lib import normalizer
 from packtools.sps.utils import xml_utils
