+++ packtools/sps/sps_maker.py
@@ -1,6 +1,6 @@
 from packtools.sps.models import packages, sps_package
 from packtools.sps.libs import async_download, reqs
-from lib import file_utils, file_utils_mimetype
+from packtools.lib import file_utils, file_utils_mimetype
 from packtools.sps import exceptions
