+++ tests/sps/models/test_sps_package.py
@@ -1,4 +1,4 @@
-from lib.normalizer import extract_number_and_supplment_from_issue_element
+from packtools.lib.normalizer import extract_number_and_supplment_from_issue_element
 from packtools.sps.models.sps_package import (
     SPS_Package,
     Identity,
